+++ packages/cli/plugin-garfish/src/cli/index.ts
@@ -1,76 +1,84 @@
 import path from 'path';
-import {
-  createRuntimeExportsUtils,
-  getEntryOptions,
-  PLUGIN_SCHEMAS,
-} from '@modern-js/utils';
+import { createRuntimeExportsUtils, INTERNAL_PLUGINS } from '@modern-js/utils';
 import {
   createPlugin,
-  NormalizedConfig,
   useAppContext,
   useResolvedConfigContext,
 } from '@modern-js/core';
 import type WebpackChain from 'webpack-chain';
 import { logger } from '../util';
-import { makeProvider, makeRenderFunction } from './utils';
+import {
+  getRuntimeConfig,
+  makeProvider,
+  makeRenderFunction,
+  setRuntimeConfig,
+} from './utils';
 
-function getRuntimeConfig(config: NormalizedConfig) {
-  if (config?.runtime?.feature) {
-    return config?.runtime?.feature;
-  }
-  return config?.runtime;
-}
+export const externals = { 'react-dom': 'react-dom', react: 'react' };
 
-function setRuntimeConfig(config: NormalizedConfig, key: string, value: any) {
-  if (config?.runtime && config?.runtime[key]) {
-    config.runtime[key] = value;
-  }
-  if (config?.runtime?.feature && config?.runtime?.feature[key]) {
-    config.runtime.feature[key] = value;
-  }
-}
+type Initializer = Parameters<typeof createPlugin>[0];
+// eslint-disable-next-line @typescript-eslint/no-invalid-void-type
+type NonVoidPromiseAble<T> = T extends void | Promise<any> ? never : T;
+export type LifeCycle = NonVoidPromiseAble<ReturnType<Initializer>>;
 
-const externals = { 'react-dom': 'react-dom', react: 'react' };
-export default createPlugin(
+export const resolvedConfig: NonNullable<
+  LifeCycle['resolvedConfig']
+> = async config => {
+  const { resolved } = config;
+  const { masterApp, router } = getRuntimeConfig(resolved);
+
+  const nConfig = {
+    resolved: {
+      ...resolved,
+    },
+  };
+
+  if (masterApp) {
+    // basename does not exist use router's basename
+    setRuntimeConfig(
+      nConfig.resolved,
+      'masterApp',
+      Object.assign(typeof masterApp === 'object' ? { ...masterApp } : {}, {
+        basename: router?.historyOptions?.basename || '/',
+      }),
+    );
+  }
+
+  logger(`resolvedConfig`, {
+    runtime: nConfig.resolved.runtime,
+    deploy: nConfig.resolved.deploy,
+    server: nConfig.resolved.server,
+  });
+  return nConfig;
+};
+
+export const initializer: (
+  hooks: {
+    resolvedConfig: LifeCycle['resolvedConfig'];
+    validateSchema: LifeCycle['validateSchema'];
+  },
+  initializerConfig: {
+    runtimePluginName?: string;
+    defaultEnableHtmlEntry?: boolean;
+    defaultExternalBasicLibrary?: boolean;
+  },
+) => Initializer =
+  (
+    // eslint-disable-next-line @typescript-eslint/no-shadow
+    { resolvedConfig, validateSchema },
+    {
+      runtimePluginName = '@modern-js/runtime/plugins',
+      defaultEnableHtmlEntry = true,
+      defaultExternalBasicLibrary = false,
+    },
+  ) =>
   () => {
-    const configMap = new Map<string, any>();
     let pluginsExportsUtils: ReturnType<typeof createRuntimeExportsUtils>;
-
     let runtimeExportsUtils: ReturnType<typeof createRuntimeExportsUtils>;
 
     return {
-      resolvedConfig(config) {
-        const { resolved } = config;
-        const { masterApp, router } = getRuntimeConfig(resolved);
-
-        const nConfig = {
-          resolved: {
-            ...resolved,
-          },
-        };
-
-        if (masterApp) {
-          // basename does not exist use router's basename
-          setRuntimeConfig(
-            nConfig.resolved,
-            'masterApp',
-            Object.assign(
-              typeof masterApp === 'object' ? { ...masterApp } : {},
-              {
-                basename: router?.basename || '/',
-              },
-            ),
-          );
-        }
-
-        logger(`resolvedConfig`, {
-          runtime: nConfig.resolved.runtime,
-          deploy: nConfig.resolved.deploy,
-          server: nConfig.resolved.server,
-        });
-
-        return nConfig;
-      },
+      validateSchema,
+      resolvedConfig,
       config() {
         // eslint-disable-next-line react-hooks/rules-of-hooks
         const config = useAppContext();
@@ -97,31 +105,32 @@ export default createPlugin(
               {
                 chain,
                 webpack,
+                env = process.env.NODE_ENV || 'development',
               }: { chain: WebpackChain; webpack: any; env: string },
             ) => {
-              const env = process.env.NODE_ENV;
               // eslint-disable-next-line react-hooks/rules-of-hooks
-              const resolvedConfig = useResolvedConfigContext();
-
-              if (resolvedConfig?.deploy?.microFrontend) {
+              const resolveOptions = useResolvedConfigContext();
+              if (resolveOptions?.deploy?.microFrontend) {
                 chain.output.libraryTarget('umd');
-
-                if (resolvedConfig.server.port) {
+                if (resolveOptions?.server?.port) {
                   chain.output.publicPath(
                     env === 'development'
-                      ? `//localhost:${resolvedConfig.server.port}/`
+                      ? `//localhost:${resolveOptions.server.port}/`
                       : webpackConfig.output.publicPath,
                   );
                 }
-
                 // add comments avoid sourcemap abnormal
-                chain
-                  .plugin('banner')
-                  .use(webpack.BannerPlugin, [{ banner: 'Micro front-end' }]);
-
-                const { enableHtmlEntry = true, externalBasicLibrary = true } =
-                  typeof resolvedConfig?.deploy?.microFrontend === 'object'
-                    ? resolvedConfig?.deploy?.microFrontend
+                if (webpack.BannerPlugin) {
+                  chain
+                    .plugin('banner')
+                    .use(webpack.BannerPlugin, [{ banner: 'Micro front-end' }]);
+                }
+                const {
+                  enableHtmlEntry = defaultEnableHtmlEntry,
+                  externalBasicLibrary = defaultExternalBasicLibrary,
+                } =
+                  typeof resolveOptions?.deploy?.microFrontend === 'object'
+                    ? resolveOptions?.deploy?.microFrontend
                     : {};
                 // external
                 if (externalBasicLibrary) {
@@ -137,7 +146,6 @@ export default createPlugin(
                   });
                 }
               }
-
               const resolveWebpackConfig = chain.toConfig();
               logger('webpackConfig', {
                 output: resolveWebpackConfig.output,
@@ -151,38 +159,31 @@ export default createPlugin(
       },
       addRuntimeExports() {
         const mfPackage = path.resolve(__dirname, '../../../../');
-        const addExportStatement = `export { default as garfish } from '${mfPackage}'`;
+        const addExportStatement = `export { default as garfish, default as masterApp } from '${mfPackage}'`;
         logger('exportStatement', addExportStatement);
         pluginsExportsUtils.addExport(addExportStatement);
 
         runtimeExportsUtils.addExport(`export * from '${mfPackage}'`);
       },
-      validateSchema() {
-        return PLUGIN_SCHEMAS['@modern-js/plugin-garfish'];
-      },
       modifyEntryImports({ entrypoint, imports }: any) {
         // eslint-disable-next-line react-hooks/rules-of-hooks
-        const resolvedConfig = useResolvedConfigContext();
-        // support legacy config
-        const { masterApp } = getRuntimeConfig(resolvedConfig);
-        // eslint-disable-next-line react-hooks/rules-of-hooks
-        const { packageName } = useAppContext();
+        const config = useResolvedConfigContext();
+        const { masterApp } = getRuntimeConfig(config);
 
-        const masterAppConfig = getEntryOptions(
-          entrypoint.entryName,
-          masterApp,
-          resolvedConfig.runtimeByEntries,
-          packageName,
-        );
-
-        configMap.set(entrypoint.entryName, masterAppConfig);
-
-        if (masterAppConfig) {
+        if (masterApp) {
           imports.push({
-            value: '@modern-js/runtime/plugins',
+            value: runtimePluginName,
             specifiers: [
               {
                 imported: 'garfish',
+              },
+            ],
+          });
+          imports.push({
+            value: runtimePluginName,
+            specifiers: [
+              {
+                imported: 'masterApp',
               },
             ],
           });
@@ -203,15 +204,20 @@ export default createPlugin(
         return { imports, entrypoint };
       },
       modifyEntryRuntimePlugins({ entrypoint, plugins }: any) {
-        const masterAppConfig = configMap.get(entrypoint.entryName);
+        // eslint-disable-next-line react-hooks/rules-of-hooks
+        const config = useResolvedConfigContext();
+        const { masterApp } = getRuntimeConfig(config);
 
-        if (masterAppConfig) {
-          logger('garfishPlugin options', masterAppConfig);
+        if (masterApp) {
+          logger('garfishPlugin options', masterApp);
 
           plugins.push({
             name: 'garfish',
             args: 'masterApp',
-            options: JSON.stringify(masterAppConfig),
+            options:
+              masterApp === true
+                ? JSON.stringify({})
+                : JSON.stringify(masterApp),
           });
         }
         return { entrypoint, plugins };
@@ -232,8 +238,8 @@ export default createPlugin(
       },
       modifyEntryExport({ entrypoint, exportStatement }: any) {
         // eslint-disable-next-line react-hooks/rules-of-hooks
-        const resolvedConfig = useResolvedConfigContext();
-        if (resolvedConfig?.deploy?.microFrontend) {
+        const config = useResolvedConfigContext();
+        if (config?.deploy?.microFrontend) {
           const exportStatementCode = makeProvider();
           logger('exportStatement', exportStatementCode);
           return {
@@ -248,7 +254,20 @@ export default createPlugin(
         };
       },
     };
-  },
+  };
+
+export default createPlugin(
+  initializer(
+    {
+      resolvedConfig,
+      validateSchema() {
+        return INTERNAL_PLUGINS['@modern-js/plugin-garfish'];
+      },
+    },
+    {
+      runtimePluginName: '@modern-js/runtime/plugins',
+    },
+  ),
   {
     name: '@modern-js/plugin-garfish',
   },