+++ packages/cli/plugin-unbundle/src/constants.ts
@@ -95,7 +95,11 @@ export const LAMBDA_API_FUNCTION_QUERY = '__lambda_api_function__';
 export const IGNORE_OPTIMIZE_DPES = ['domino'];
 
 // default lazy import ui components library
-export const DEFAULT_LAZY_IMPORT_UI_COMPONENTS = ['antd'];
+export const DEFAULT_LAZY_IMPORT_UI_COMPONENTS = [
+  'antd',
+  '@arco-design/web-react',
+  '@arco-design/web-react/icon',
+];
 
 export const ESBUILD_RESOLVE_PLUGIN_NAME = 'esm-resolve-plugin';