+++ packages/generator/generator-common/src/newAction/module/index.ts
@@ -75,11 +75,17 @@ export const ModuleActionFunctionsDependencies: Partial<
   [ActionFunction.Sass]: '@modern-js/plugin-sass',
 };
 
+export const ModuleActionFunctionsAppendTypeContent: Partial<
+  Record<ActionFunction, string>
+> = {
+  [ActionFunction.TailwindCSS]: `/// <reference types='@modern-js/plugin-tailwindcss/types' />`,
+};
+
 export const ModuleNewActionGenerators: Partial<
   Record<ActionType, Record<string, string>>
 > = {
   [ActionType.Function]: {
-    [ActionFunction.TailwindCSS]: '@modern-js/dependence-generator',
+    [ActionFunction.TailwindCSS]: '@modern-js/tailwindcss-generator',
     [ActionFunction.Less]: '@modern-js/dependence-generator',
     [ActionFunction.Sass]: '@modern-js/dependence-generator',
     [ActionFunction.I18n]: '@modern-js/dependence-generator',