+++ packages/toolkit/types/cli/index.d.ts
@@ -57,9 +57,9 @@ export interface IAppContext {
   internalDirectory: string;
   plugins: {
     cli?: any;
-    cliPath?: any;
+    cliPkg?: any;
     server?: any;
-    serverPath?: any;
+    serverPkg?: any;
   }[];
   entrypoints: Entrypoint[];
   checkedEntries: string[];
@@ -96,6 +96,10 @@ export interface Hooks {
     unknown
   >;
   afterBuild: AsyncWorkflow<void, unknown>;
+  afterMonorepoDeploy: AsyncWorkflow<
+    { operator: any; deployProjectNames: string[] },
+    void
+  >;
   beforeDeploy: AsyncWorkflow<Record<string, any>, unknown>;
   afterDeploy: AsyncWorkflow<Record<string, any>, unknown>;
   modifyEntryExport: AsyncWaterfall<{