+++ packages/toolkit/load-config/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/load-config
 
+## 1.2.2
+
+### Patch Changes
+
+- 55e18278: chore: remove unused dependencies and devDependencies
+- Updated dependencies [4c792f68]
+- Updated dependencies [a7f42f48]
+  - @modern-js/utils@1.3.3
+
 ## 1.2.1
 
 ### Patch Changes